+++ NEWS.md
@@ -1,7 +1,8 @@
 # Wavelength 0.0.0.9000
 
+* Added function to import Burundi data, `structure_bdi()`
 * Added function to pull out header row names, `extract_headers()`
-* Added function to import Hait dataset, `structure_hti()`
+* Added function to import Haiti dataset, `structure_hti()`
 * Added function for reshaping long, `reshape_long()`
 * Added function for importing Haiti that pulls in and combines multiple header rows, `identify_headers_hti()`
 * Added function to import Uganda system output, `structure_tza()`